add information about how much disk space is sufficient.
I based this on looking at the production webapp node and saw that it was created with 5gb and has 37% used, and a couchdb node that was allocated 10gb and had 16% used, and an openvpn node which was created with 12gb and had 11% used. So that seemed to me to indicate that 10gb would be sufficient for an initial start and then some growth.